// src/enums/sectionsNames.ts
export enum SectionsNames {
  Frontend = "Frontend",
  MarkdownLanguages = "Markdown Languages",
  ProgrammingLanguages = "Programming languages",
  JavascriptLibraries = "Javascript libraries and frameworks",
  MobileFrameworks = "Mobile frameworks",
  PhpFrameworks = "PHP frameworks",
  JavaFrameworks = "Java frameworks",
  ReactNativeLibraries = "React Native Libraries",
  FormLibraries = "Form libraries",
  StateManagementLibraries = "State management libraries",
  FetchingLibraries = "Fetching libraries",
  UiLibraries = "UI libraries",
  DataVisualization = "Data visualization",
  AnimationsLibraries = "Animations libraries",
  TestingFrameworksAndTools = "Testing frameworks and tools",
  Internationalization = "Internationalization",
  Maps = "Maps",
  BackendTechnologies = "Backend",
  Cloud = "Cloud",
  Databases = "Databases",
  Cms = "CMS",
  PaymentTools = "Payment tools",
  WebServices = "Web services",
  AuthorizationTools = "Authorization tools",
  Containerization = "Containerization",
  Web3 = "Web 3",
  Analytics = "Analytics",
  Security = "Security",
  CiCd = "CI CD",
  SpeechRecognition = "Speech recognition",
  Virtualization = "Virtualization",
  Documentation = "Documentation",
  LoggingTools = "Logging tools",
  PerformanceMonitoringTools = "Performance monitoring tools",
  ModuleBundlers = "Module bundlers",
  GraphicEditors = "Graphic editors",
  VersionControlSystems = "Version control systems",
  AITools = "AI tools",
  NotFound = "Not found",
}